add single function for str-strip-chars/-end/-start
